+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- Add an opt-in setting to disable the submit button until the current page passes validation. ([#1514](https://github.com/verbb/formie/issues/1514))
+
 ### Fixed
 - Fix dropdown field options not saving in the form builder when no default option is selected. ([#2795](https://github.com/verbb/formie/issues/2795))
 - Fix replacing a deleted field with a new field using the same handle requiring a two-step save in the form builder. ([#1854](https://github.com/verbb/formie/issues/1854))